Schema Definition:
const userSchema = new Schema({ ... }, { timestamps: true });: This line defines a schema for the User model. A schema is a structure that defines the shape and content of documents within a MongoDB collection. The schema includes three fields:
name: A string field to store the user's name.
email: A string field to store the user's email address.
password: A string field to store the user's hashed password.
The schema also has an option { timestamps: true } which tells Mongoose to automatically manage createdAt and updatedAt fields for each document in the collection.