Model Creation:
const User = mongoose.models.User || mongoose.model("User", userSchema);: This line defines the User model based on the userSchema. It first checks if a model named User already exists in mongoose.models to avoid redefining the model. If it exists, it uses the existing model; otherwise, it creates a new model. Models in Mongoose are constructors compiled from schemas and are used to create and read documents from the underlying MongoDB database.