Session Configuration:
required: true: This option specifies that authentication is required for this component. If the user is not authenticated, the onUnauthenticated callback will be triggered.
onUnauthenticated() { redirect("/api/auth/signin?callbackUrl=/ClientMember"); }: This callback function is executed if the user is not authenticated. It redirects the user to the sign-in page, passing a callback URL that navigates back to /ClientMember after a successful sign-in.