Rendering Navigation Bar:
The component returns JSX, which describes the structure of the navigation bar.
The navigation bar is wrapped in a <header> element with a background color of gray (bg-gray-600) and white text color (text-gray-100).
Inside the header, there's a <nav> element with flexbox properties to align its children.
The navigation links are placed inside a <div> element with the class flex gap-10, ensuring spacing between the links.
Navigation Links:

There are several <Link> components imported from Next.js, each representing a navigation link to a different page of the website.
The links include "Home", "Create User", "Client Member", "Member", and "Public" pages.
Conditional Rendering for Authentication:

The login/logout link is conditionally rendered based on the existence of the session object.
If session exists (meaning the user is authenticated), a logout link is displayed, which upon clicking, logs the user out and redirects to the home page (/).
If session doesn't exist (meaning the user is not authenticated), a login link is displayed, which upon clicking, redirects the user to the sign-in page (/api/auth/signin).