Logging and Authorization Check:
console.log(req.nextUrl.pathname);: This logs the path of the requested URL.
console.log(req.nextauth.token.role);: This logs the user's role extracted from the authentication token.
The middleware then checks if the requested path starts with "/CreateUser" and if the user's role is not "admin".
If these conditions are met, it rewrites the request URL to redirect to the "/Denied" page.